requirements.txt
# Cloud Provider SDKs
boto3>=1.26.0
azure-identity>=1.12.0
azure-mgmt-resource>=21.1.0
azure-mgmt-compute>=29.1.0
azure-mgmt-network>=21.0.1
azure-mgmt-storage>=20.1.0
azure-mgmt-monitor>=5.0.1
azure-mgmt-sql>=3.0.1
azure-mgmt-subscription>=3.1.1
google-cloud-storage>=2.7.0
google-cloud-compute>=1.11.0
google-cloud-iam>=2.11.0
google-cloud-monitoring>=2.14.1
google-cloud-core>=2.3.2
google-api-python-client>=2.0.0
google-cloud-iam-v1>=0.1.0

# Database Support
pymongo>=4.6.0

# CLI and Configuration
click>=8.1.3
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Reporting
pandas>=2.0.0
tabulate>=0.9.0
rich>=13.3.5

# Utilities
requests>=2.28.2
typing-extensions>=4.5.0
python-dateutil>=2.8.2 